requirements.txt
# Crawling
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
